191: refactor home page and post
